Support color themes
Support the full range of configurable colors in a Windows Classic
theme. The colors are configured using CSS variables named after the
corresponding Windows theme file property converted from PascalCase to
kebab-case.

Colors are given in rgb format to make it easier to convert and
recognize values from a Windows theme file.

The old variables are kept for backwards compatibility and are mapped
accordingly. However, the old `--button-face`, which maps to
`--button-light`, has not been mapped because of a name conflict with
the new `--button-face`; this is a breaking change.