Fix issue where active button style pushes content down 1px
Previously, when a button was pressed, the content below it would shift
down 1px. This bug was introduced in #44 which added the effect of the
text shifting left and down 1px on press.

After this commit, the text of the button still shifts down and left 1px
but the overall height of the button does not change.

This commit introduces a small issue where the button does not shift
down visually in Safari.

Fix #145